# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * POWAMOV Infrastructure Intelligence API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type NodeStatus = (typeof NodeStatus)[keyof typeof NodeStatus];

export const NodeStatus = {
  online: "online",
  offline: "offline",
  maintenance: "maintenance",
  warning: "warning",
} as const;

export interface Node {
  id: string;
  name: string;
  location: string;
  latitude: number;
  longitude: number;
  status: NodeStatus;
  installDate: string;
  degradationPercent: number;
  totalEnergyKwh: number;
  lastMaintenanceDate?: string | null;
  vehiclePassesToday: number;
  compressionCycles: number;
}

export interface CreateNodeBody {
  name: string;
  location: string;
  latitude: number;
  longitude: number;
}

export interface TelemetryReading {
  id: string;
  nodeId: string;
  timestamp: string;
  compressionForceN: number;
  energyGeneratedWh: number;
  vehicleSpeed: number;
  vehicleWeight: number;
  temperature: number;
  vibration: number;
}

export interface NodeLiveStatus {
  nodeId: string;
  powerOutputW: number;
  recentPasses: number;
  compressionForceN: number;
  status: string;
}

export interface LiveTelemetrySnapshot {
  timestamp: string;
  totalNodesOnline: number;
  totalNodesOffline: number;
  totalNodesWarning: number;
  currentPowerOutputW: number;
  todayEnergyKwh: number;
  todayVehiclePasses: number;
  avgCompressionForceN: number;
  nodes: NodeLiveStatus[];
}

export interface TelemetryAggregate {
  timestamp: string;
  avgPowerW: number;
  totalEnergyKwh: number;
  totalPasses: number;
  avgEfficiencyPercent: number;
}

export type MaintenanceForecastUrgency =
  (typeof MaintenanceForecastUrgency)[keyof typeof MaintenanceForecastUrgency];

export const MaintenanceForecastUrgency = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface MaintenanceForecast {
  nodeId: string;
  nodeName: string;
  forecastDate: string;
  urgency: MaintenanceForecastUrgency;
  degradationRate: number;
  estimatedLifeRemainingDays: number;
  recommendedAction: string;
  estimatedCost: number;
}

export type MaintenanceAlertSeverity =
  (typeof MaintenanceAlertSeverity)[keyof typeof MaintenanceAlertSeverity];

export const MaintenanceAlertSeverity = {
  info: "info",
  warning: "warning",
  critical: "critical",
} as const;

export interface MaintenanceAlert {
  id: string;
  nodeId: string;
  nodeName: string;
  alertType: string;
  severity: MaintenanceAlertSeverity;
  message: string;
  timestamp: string;
  acknowledged: boolean;
}

export interface AnalyticsSummary {
  totalEnergyHarvestedKwh: number;
  carbonOffsetKg: number;
  gridDisplacementPercent: number;
  totalVehiclePasses: number;
  avgEfficiencyPercent: number;
  activeNodes: number;
  totalNodes: number;
  revenueUsd: number;
  co2SavedTons: number;
  treesEquivalent: number;
}

export interface EnergyHistoryPoint {
  date: string;
  energyKwh: number;
  vehiclePasses: number;
  carbonOffsetKg: number;
  efficiencyPercent: number;
}

export interface SimulationParams {
  vehicleWeightKg: number;
  vehicleSpeedKmh: number;
  vehiclesPerHour: number;
  simulationHours: number;
}

export interface SimulationDataPoint {
  hour: number;
  energyKwh: number;
  powerW: number;
  passes: number;
  compressionForceN: number;
}

export interface SimulationResult {
  nodeId: string;
  totalEnergyKwh: number;
  peakPowerW: number;
  avgCompressionForceN: number;
  totalPasses: number;
  efficiencyPercent: number;
  degradationAdded: number;
  dataPoints: SimulationDataPoint[];
}

export type GetNodeTelemetryParams = {
  limit?: number;
};

export type GetTelemetryHistoryParams = {
  hours?: number;
};

export type GetEnergyHistoryParams = {
  days?: number;
};