# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * POWAMOV Infrastructure Intelligence API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all POWAMOV nodes
 */
export const ListNodesResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  location: zod.string(),
  latitude: zod.number(),
  longitude: zod.number(),
  status: zod.enum(["online", "offline", "maintenance", "warning"]),
  installDate: zod.string(),
  degradationPercent: zod.number(),
  totalEnergyKwh: zod.number(),
  lastMaintenanceDate: zod.string().nullish(),
  vehiclePassesToday: zod.number(),
  compressionCycles: zod.number(),
});
export const ListNodesResponse = zod.array(ListNodesResponseItem);

/**
 * @summary Create a new node
 */
export const CreateNodeBody = zod.object({
  name: zod.string(),
  location: zod.string(),
  latitude: zod.number(),
  longitude: zod.number(),
});

/**
 * @summary Get a node by ID
 */
export const GetNodeParams = zod.object({
  id: zod.coerce.string(),
});

export const GetNodeResponse = zod.object({
  id: zod.string(),
  name: zod.string(),
  location: zod.string(),
  latitude: zod.number(),
  longitude: zod.number(),
  status: zod.enum(["online", "offline", "maintenance", "warning"]),
  installDate: zod.string(),
  degradationPercent: zod.number(),
  totalEnergyKwh: zod.number(),
  lastMaintenanceDate: zod.string().nullish(),
  vehiclePassesToday: zod.number(),
  compressionCycles: zod.number(),
});

/**
 * @summary Get recent telemetry for a node
 */
export const GetNodeTelemetryParams = zod.object({
  id: zod.coerce.string(),
});

export const getNodeTelemetryQueryLimitDefault = 50;

export const GetNodeTelemetryQueryParams = zod.object({
  limit: zod.coerce.number().default(getNodeTelemetryQueryLimitDefault),
});

export const GetNodeTelemetryResponseItem = zod.object({
  id: zod.string(),
  nodeId: zod.string(),
  timestamp: zod.string(),
  compressionForceN: zod.number(),
  energyGeneratedWh: zod.number(),
  vehicleSpeed: zod.number(),
  vehicleWeight: zod.number(),
  temperature: zod.number(),
  vibration: zod.number(),
});
export const GetNodeTelemetryResponse = zod.array(GetNodeTelemetryResponseItem);

/**
 * @summary Get live telemetry snapshot for all nodes
 */
export const GetLiveTelemetryResponse = zod.object({
  timestamp: zod.string(),
  totalNodesOnline: zod.number(),
  totalNodesOffline: zod.number(),
  totalNodesWarning: zod.number(),
  currentPowerOutputW: zod.number(),
  todayEnergyKwh: zod.number(),
  todayVehiclePasses: zod.number(),
  avgCompressionForceN: zod.number(),
  nodes: zod.array(
    zod.object({
      nodeId: zod.string(),
      powerOutputW: zod.number(),
      recentPasses: zod.number(),
      compressionForceN: zod.number(),
      status: zod.string(),
    }),
  ),
});

/**
 * @summary Get aggregated telemetry history
 */
export const getTelemetryHistoryQueryHoursDefault = 24;

export const GetTelemetryHistoryQueryParams = zod.object({
  hours: zod.coerce.number().default(getTelemetryHistoryQueryHoursDefault),
});

export const GetTelemetryHistoryResponseItem = zod.object({
  timestamp: zod.string(),
  avgPowerW: zod.number(),
  totalEnergyKwh: zod.number(),
  totalPasses: zod.number(),
  avgEfficiencyPercent: zod.number(),
});
export const GetTelemetryHistoryResponse = zod.array(
  GetTelemetryHistoryResponseItem,
);

/**
 * @summary Get predictive maintenance forecasts
 */
export const GetMaintenanceForecastsResponseItem = zod.object({
  nodeId: zod.string(),
  nodeName: zod.string(),
  forecastDate: zod.string(),
  urgency: zod.enum(["low", "medium", "high", "critical"]),
  degradationRate: zod.number(),
  estimatedLifeRemainingDays: zod.number(),
  recommendedAction: zod.string(),
  estimatedCost: zod.number(),
});
export const GetMaintenanceForecastsResponse = zod.array(
  GetMaintenanceForecastsResponseItem,
);

/**
 * @summary Get active maintenance alerts
 */
export const GetMaintenanceAlertsResponseItem = zod.object({
  id: zod.string(),
  nodeId: zod.string(),
  nodeName: zod.string(),
  alertType: zod.string(),
  severity: zod.enum(["info", "warning", "critical"]),
  message: zod.string(),
  timestamp: zod.string(),
  acknowledged: zod.boolean(),
});
export const GetMaintenanceAlertsResponse = zod.array(
  GetMaintenanceAlertsResponseItem,
);

/**
 * @summary Get overall energy analytics summary
 */
export const GetAnalyticsSummaryResponse = zod.object({
  totalEnergyHarvestedKwh: zod.number(),
  carbonOffsetKg: zod.number(),
  gridDisplacementPercent: zod.number(),
  totalVehiclePasses: zod.number(),
  avgEfficiencyPercent: zod.number(),
  activeNodes: zod.number(),
  totalNodes: zod.number(),
  revenueUsd: zod.number(),
  co2SavedTons: zod.number(),
  treesEquivalent: zod.number(),
});

/**
 * @summary Get energy generation history
 */
export const getEnergyHistoryQueryDaysDefault = 30;

export const GetEnergyHistoryQueryParams = zod.object({
  days: zod.coerce.number().default(getEnergyHistoryQueryDaysDefault),
});

export const GetEnergyHistoryResponseItem = zod.object({
  date: zod.string(),
  energyKwh: zod.number(),
  vehiclePasses: zod.number(),
  carbonOffsetKg: zod.number(),
  efficiencyPercent: zod.number(),
});
export const GetEnergyHistoryResponse = zod.array(GetEnergyHistoryResponseItem);

/**
 * @summary Run digital twin compression simulation for a node
 */
export const SimulateDigitalTwinParams = zod.object({
  nodeId: zod.coerce.string(),
});

export const SimulateDigitalTwinBody = zod.object({
  vehicleWeightKg: zod.number(),
  vehicleSpeedKmh: zod.number(),
  vehiclesPerHour: zod.number(),
  simulationHours: zod.number(),
});

export const SimulateDigitalTwinResponse = zod.object({
  nodeId: zod.string(),
  totalEnergyKwh: zod.number(),
  peakPowerW: zod.number(),
  avgCompressionForceN: zod.number(),
  totalPasses: zod.number(),
  efficiencyPercent: zod.number(),
  degradationAdded: zod.number(),
  dataPoints: zod.array(
    zod.object({
      hour: zod.number(),
      energyKwh: zod.number(),
      powerW: zod.number(),
      passes: zod.number(),
      compressionForceN: zod.number(),
    }),
  ),
});