# lib/api-zod/src/generated/types/analyticsSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * POWAMOV Infrastructure Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface AnalyticsSummary {
  totalEnergyHarvestedKwh: number;
  carbonOffsetKg: number;
  gridDisplacementPercent: number;
  totalVehiclePasses: number;
  avgEfficiencyPercent: number;
  activeNodes: number;
  totalNodes: number;
  revenueUsd: number;
  co2SavedTons: number;
  treesEquivalent: number;
}