# lib/api-zod/src/generated/types/createNodeBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * POWAMOV Infrastructure Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateNodeBody {
  name: string;
  location: string;
  latitude: number;
  longitude: number;
}