# lib/api-zod/src/generated/types/maintenanceForecast.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * POWAMOV Infrastructure Intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { MaintenanceForecastUrgency } from "./maintenanceForecastUrgency";

export interface MaintenanceForecast {
  nodeId: string;
  nodeName: string;
  forecastDate: string;
  urgency: MaintenanceForecastUrgency;
  degradationRate: number;
  estimatedLifeRemainingDays: number;
  recommendedAction: string;
  estimatedCost: number;
}