# lib/api-zod/src/generated/types/node.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * POWAMOV Infrastructure Intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { NodeStatus } from "./nodeStatus";

export interface Node {
  id: string;
  name: string;
  location: string;
  latitude: number;
  longitude: number;
  status: NodeStatus;
  installDate: string;
  degradationPercent: number;
  totalEnergyKwh: number;
  lastMaintenanceDate?: string | null;
  vehiclePassesToday: number;
  compressionCycles: number;
}