# lib/api-zod/src/generated/types/simulationDataPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * POWAMOV Infrastructure Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface SimulationDataPoint {
  hour: number;
  energyKwh: number;
  powerW: number;
  passes: number;
  compressionForceN: number;
}