# lib/api-zod/src/generated/types/simulationResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * POWAMOV Infrastructure Intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { SimulationDataPoint } from "./simulationDataPoint";

export interface SimulationResult {
  nodeId: string;
  totalEnergyKwh: number;
  peakPowerW: number;
  avgCompressionForceN: number;
  totalPasses: number;
  efficiencyPercent: number;
  degradationAdded: number;
  dataPoints: SimulationDataPoint[];
}